ai-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.33.2
torch==2.0.1
requests==2.31.0
python-dotenv==1.0.0
rasa==3.6.4
google-cloud-dialogflow==2.24.1
pymongo==4.5.0
redis==4.6.0
